functions/consume-catalog/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.2.2
    # via
    #   -r requirements.in
    #   google-auth
certifi==2021.5.30
    # via
    #   -r requirements.in
    #   requests
cffi==1.14.5
    # via
    #   -r requirements.in
    #   google-crc32c
chardet==4.0.0
    # via
    #   -r requirements.in
    #   requests
ckanapi==4.6
    # via -r requirements.in
dataclasses==0.6
    # via -r requirements.in
docopt==0.6.2
    # via
    #   -r requirements.in
    #   ckanapi
google-api-core[grpc]==1.29.0
    # via
    #   -r requirements.in
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
google-auth==1.30.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
google-cloud-core==1.4.3
    # via
    #   -r requirements.in
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==1.15.1
    # via -r requirements.in
google-cloud-secret-manager==2.0.0
    # via -r requirements.in
google-cloud-storage==1.32.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via
    #   -r requirements.in
    #   google-resumable-media
google-resumable-media==1.3.0
    # via
    #   -r requirements.in
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.53.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.12.3
    # via
    #   -r requirements.in
    #   google-cloud-secret-manager
grpcio==1.38.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
idna==2.10
    # via
    #   -r requirements.in
    #   requests
libcst==0.3.19
    # via
    #   -r requirements.in
    #   google-cloud-secret-manager
mypy-extensions==0.4.3
    # via
    #   -r requirements.in
    #   typing-inspect
packaging==20.9
    # via
    #   -r requirements.in
    #   google-api-core
proto-plus==1.18.1
    # via
    #   -r requirements.in
    #   google-cloud-secret-manager
protobuf==3.17.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.4.8
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via
    #   -r requirements.in
    #   google-auth
pycparser==2.20
    # via
    #   -r requirements.in
    #   cffi
pyparsing==2.4.7
    # via
    #   -r requirements.in
    #   packaging
python-slugify==5.0.2
    # via
    #   -r requirements.in
    #   ckanapi
pytz==2021.1
    # via
    #   -r requirements.in
    #   google-api-core
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   libcst
requests==2.25.1
    # via
    #   -r requirements.in
    #   ckanapi
    #   google-api-core
    #   google-cloud-storage
rsa==4.7.2
    # via
    #   -r requirements.in
    #   google-auth
six==1.16.0
    # via
    #   -r requirements.in
    #   ckanapi
    #   google-api-core
    #   google-auth
    #   google-resumable-media
    #   grpcio
    #   protobuf
text-unidecode==1.3
    # via
    #   -r requirements.in
    #   python-slugify
typing-extensions==3.10.0.0
    # via
    #   -r requirements.in
    #   libcst
    #   typing-inspect
typing-inspect==0.7.0
    # via
    #   -r requirements.in
    #   libcst
urllib3==1.26.5
    # via
    #   -r requirements.in
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
